[Bugfix] Resource:take only works at beginning of the game
Resource:take expects there to always be a pile of resources in the supply and for that pile to always have the guid. This isn't the case because restacking the resources in a different order will change their guid. Instead of guid, we should use the location of the supply tiles to find the resources. This code uses a raycast since it can find the topmost resource in the pile even if they're not stacked